pyproject.toml
[project]
name = "SPARMVET_VIZ"
version = "0.1.0"
dependencies = [
    "polars>=1.0.0",
    "shiny>=1.0.0",
    "numpy>=2.0.0",
    "pyyaml>=6.0.0",
    "plotnine>=0.14.0",
    "fastexcel>=0.11.0", # Required for ultra-fast Excel parsing in Polars
]

# [tool.ruff]
# # The Architect's "Standard of Excellence"
# line-length = 100
# select = ["E", "F", "I", "N"] # Errors, Flakes, Isort (Import sorting), Naming

# [tool.ruff.isort]
# known-first-party = ["data_prep", "viz_core"]
